+++ development/esm-jest-transformer.js
@@ -1,7 +1,0 @@
-const { transform } = require('@babel/core');
-
-module.exports = {
-  process(src) {
-    return `require = require('esm')(module);\n${src}`;
-  },
-};