+++ src/subscribe/index.js
@@ -1,2 +1,2 @@
-export { default as PendingTransaction } from './PendingTransaction';
-export { default as LogIterator } from './LogIterator';
+module.exports.PendingTransaction = require('./PendingTransaction');
+module.exports.LogIterator = require('./LogIterator');