+++ CHANGE_LOG.md
@@ -1,5 +1,27 @@
 # change log
 
+## v0.10.0-alpha
+
+* add `getStatus`
+
+```
+cfx.getStatus()
+```
+
+* remove `getRiskCoefficient` and replace with `getConfirmationRiskByHash`
+
+```
+// old
+cfx.getRiskCoefficient(epochNumber)
+
+// new
+cfx.getConfirmationRiskByHash(blockHash)
+```
+
+* remove `getAccount` cause it's internal RPC.
+
+* use `require` replace `import` to gen code.
+
 ## v0.9.2
 
 * add defaultStorageLimit and defaultChainId for Conflux